requirements.txt
# Web Framework
Flask>=2.3.0

# Computer Vision & Face Recognition
opencv-python>=4.8.0
numpy>=1.24.0
mtcnn>=0.1.1
keras-facenet>=0.3.0
tensorflow>=2.13.0

# MQTT & Redis
paho-mqtt>=1.6.0
redis>=4.5.0

# Utilities
Pillow>=9.5.0

